requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
nltk==3.8.1
pydantic==2.6.1
python-multipart==0.0.9
requests==2.31.0
pytest==7.4.3
httpx==0.26.0
prometheus-client==0.19.0
psutil==5.9.8
spacy==3.7.2
transformers==4.37.2
torch==2.2.0
numpy==1.26.2
pytest-cov==4.1.0
networkx==3.2.1
flask-sqlalchemy==3.1.1
alembic==1.13.1 